Fix another compilation warning
svn path=/trunk/KDE/kdenetwork/kopete/; revision=1127140

+++ protocols/oscar/liboscar/inputprotocolbase.cpp
@@ -87,7 +87,7 @@ bool InputProtocolBase::safeReadBytes( QByteArray& data, uint & len )
 		// so look for that in the last position instead of \0
 		// this caused a crash - guessing that temp.length() is set to the number of bytes actually read...
 		// if ( (quint8)( * ( temp.data() + ( temp.length() - 1 ) ) ) == 0xFF )
-		if ( temp.length() < ( val - 1 ) )
+		if ( (uint)temp.length() < ( val - 1 ) )
 		{
 			qDebug( "InputProtocol::safeReadBytes() - string broke, giving up, only got: %i bytes out of %i",  temp.length(), val );
 			m_state = NeedMore;
